Sword Animation implemented
