+++ Assets/Items/Melee Weapons.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9f307eda39dee8c41b4a6f645b2dc1cb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
