+++ Assets/Items/Melee Weapons/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4c67b17c42151a54f95184e3576758c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
