+++ Assets/Items/Melee Weapons/Animation/Sword.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c361a4b5684062a4ba35e3e1ca6fbbfc
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
