+++ Assets/Items/Melee Weapons/Animation/SwordIdle.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ee0154b544494d9479995be373b953fb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
