+++ Assets/Items/Melee Weapons/Animation/SwordSwing.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8fd55229d9165fa46b9a3e7c115fda74
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
