+++ Assets/Items/Melee Weapons/Sprites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bc5daa84e16322d46a4edd84bb7c39f4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
