+++ Assets/Items/Melee Weapons/Sword (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9d6c127f01a55ed40989cd5449e03790
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
